services/dictionary-service/pyproject.toml
[project]
name = "dictionary-service"
version = "0.1.0"
description = "A smart dictionary and linguistic service"
readme = "README.md"
authors = [
    { name = "shubh305", email = "shubham.305singh@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "spacy>=3.7.0",
    "textblob>=0.18.0",
    "inflect>=7.0.0",
    "language-tool-python>=2.7.1",
    "googletrans-py>=4.0.0rc1",
    "pydantic-settings>=2.4.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "opentelemetry-api>=1.26.0",
    "opentelemetry-sdk>=1.26.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0",
    "redis>=5.0.0",
    "fastapi-limiter==0.1.6",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "langdetect>=1.0.9",
]

[project.scripts]
dictionary-service = "dictionary_service.main:app"

[build-system]
requires = ["uv_build>=0.1.0"]
build-backend = "uv_build"
